% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runapp.R
\name{runRasToFrameApp}
\alias{runRasToFrameApp}
\title{Start the Survey Solutions Raster-to-Frame Application}
\usage{
runRasToFrameApp(launch.browser = T)
}
\arguments{
\item{launch.browser}{If true, the system's default web browser will be
launched automatically after the app is started. Defaults to true in
interactive sessions only. The value of this parameter can also be a
function to call with the application's URL.}
}
\description{
A simple shiny application which allows to update a spatial polygons area sampling frame with one ore more
raster layers, which can be used for PPS sampling in the susospatsample app. Moreover, the application allows
also for the creation of a composite measure of size, using several input raster for a single MOS.
}
\details{
This application is part of the large set of tools, to facilitate survey implementation with
\href{https://docs.mysurvey.solutions/}{Survey Solutions}. The application is based on the
\href{https://cran.r-project.org/web/packages/terra/index.html}{terra} and the
\href{https://cran.r-project.org/web/packages/sf/index.html}{sf} package.
}
\examples{
\dontrun{
# Start app in the current working directory
runApp()

# Start app in a subdirectory called myapp
runApp("myapp")
}

## Only run this example in interactive R sessions
if (interactive()) {
  options(device.ask.default = FALSE)

  # Apps can be run without a server.r and ui.r file
  runApp(list(
    ui = bootstrapPage(
      numericInput('n', 'Number of obs', 100),
      plotOutput('plot')
    ),
    server = function(input, output) {
      output$plot <- renderPlot({ hist(runif(input$n)) })
    }
  ))


  # Running a Shiny app object
  app <- shinyApp(
    ui = bootstrapPage(
      numericInput('n', 'Number of obs', 100),
      plotOutput('plot')
    ),
    server = function(input, output) {
      output$plot <- renderPlot({ hist(runif(input$n)) })
    }
  )
  runApp(app)
}
}
