% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susorastoframe-package.R
\docType{package}
\name{susorastoframe-package}
\alias{susorastoframe}
\alias{susorastoframe-package}
\title{susorastoframe: Shiny Application to aggregate raster data within polygons for PPS sampling}
\description{
A simple shiny application which allows to update a spatial polygons area sampling frame with one ore more raster layers, which can be used for PPS sampling in the susospatsample app. Moreover, the application allows also for the creation of a composite measure of size, using several input raster for a single MOS. The application is based on the terra and the sf package.
}
\author{
\strong{Maintainer}: Michael Wild \email{mwild@worldbank.org}

}
\keyword{internal}
